カートに商品がなかったとき

+++ MISKM/Foodies/php/Cart.php
@@ -11,7 +11,11 @@
     <?php
     /* データベース接続 */
     if (isset($_SESSION['User'])) {
-      if(!empty($_SESSION['Cart'])){
+        $sql = $pdo -> prepare('select * from Cart where user_id = ?');
+        $sql -> execute([
+            $_SESSION['User']['id']
+        ]);
+        $setid = fetchAll(PDO::FETCH_ASSOC);
         $userId = $_SESSION['User']['id'];
         $pdo = new PDO($connect, USER, PASS);
         $sql = "SELECT Product.id, Product.name, Product.size, Product.price, Product.image, Cart.buy_counts
@@ -23,24 +27,25 @@
         $stmt->execute();
         $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
         /* 商品一覧 */
-        foreach ($result as $row) {
-            $id = $row['id'];
-            echo '<div class="main">';
-            echo '<figure class="image">';
-            echo '<img src="/MISKM/img/', $row['image'], '" class="cart_img">';
-            echo '</figure>';
-            echo '<div class="item">';
-            echo "<p class='name'>{$row['name']}</p>";
-            echo "<br>";
-            echo "<p class='text'>size:{$row['size']}　￥{$row['price']}<br>counts{$row['buy_counts']}</p>";
-            echo '<a href="Cart_delete.php?id=', $id, '">削除</a>';
-            echo '</div>'; // .item divを閉じる
-            echo '</div>'; // .main divを閉じる
-        }
-        echo '<button type="button" onclick="location.href=\'O_pro.php\'">購入手続きへ</button>';
-      }else{
-         echo '<p class ="error">カートに商品が入っていません。</p>';
-      }
+        if(!empty($setid)){
+            foreach ($result as $row) {
+                $id = $row['id'];
+                echo '<div class="main">';
+                echo '<figure class="image">';
+                echo '<img src="/MISKM/img/', $row['image'], '" class="cart_img">';
+                echo '</figure>';
+                echo '<div class="item">';
+                echo "<p class='name'>{$row['name']}</p>";
+                echo "<br>";
+                echo "<p class='text'>size:{$row['size']}　￥{$row['price']}<br>counts{$row['buy_counts']}</p>";
+                echo '<a href="Cart_delete.php?id=', $id, '">削除</a>';
+                echo '</div>'; // .item divを閉じる
+                echo '</div>'; // .main divを閉じる
+            }
+            echo '<button type="button" onclick="location.href=\'O_pro.php\'">購入手続きへ</button>';
+        }else{
+            echo '<p class="error">カートに商品が入っていません。</p>';
+        }    
     }else{
         echo  '<p class ="error">ログインしてください</p>';
     }
